+++ packages/amplify-graphql-api-construct/src/__tests__/__functional__/relationships.test.ts
@@ -1,16 +1,17 @@
 import * as cdk from 'aws-cdk-lib';
 import { Template } from 'aws-cdk-lib/assertions';
 import { AmplifyGraphqlApi } from '../../amplify-graphql-api';
+import { AmplifyGraphqlDefinition } from '../../amplify-graphql-definition';
 
 /**
  * Utility to wrap construct creation a basic synth step to smoke test
- * @param buildApp callback to create the resources in the stack
+ * @param schema schema to synthesize
  */
 const verifySchema = (schema: string): void => {
   const stack = new cdk.Stack();
   new AmplifyGraphqlApi(stack, 'TestApi', {
-    schema,
-    authorizationConfig: {
+    definition: AmplifyGraphqlDefinition.fromString(schema),
+    authorizationModes: {
       apiKeyConfig: { expires: cdk.Duration.days(7) },
     },
   });