+++ packages/amplify-graphql-api-construct/src/internal/default-parameters.ts
@@ -1,11 +1,13 @@
-import { SchemaTranslationBehavior } from '../types';
+import { TranslationBehavior } from '../types';
 
 /**
  * Defaults which will be used by the construct if overrides are not provided.
+ * These should not really be updated post-launch, since they're likely breaking changes for customers.
+ * Be sure to document default values when adding new keys in the top level `types` file.
  */
-export const defaultSchemaTranslationBehavior: SchemaTranslationBehavior = {
+export const defaultTranslationBehavior: TranslationBehavior = {
   shouldDeepMergeDirectiveConfigDefaults: true,
-  disableResolverDeduping: false,
+  disableResolverDeduping: true,
   sandboxModeEnabled: false,
   useSubUsernameForDefaultIdentityClaim: true,
   populateOwnerFieldForStaticGroupAuth: true,
@@ -15,4 +17,5 @@ export const defaultSchemaTranslationBehavior: SchemaTranslationBehavior = {
   respectPrimaryKeyAttributesOnConnectionField: true,
   enableSearchNodeToNodeEncryption: false,
   useAmplifyManagedTableResources: false, // TODO: flip to true once all implementations are finished
+  enableTransformerCfnOutputs: false,
 };